more creatures, things

+++ src/main/scala/Brick.scala
@@ -44,7 +44,7 @@ class Brick(name:String, world:World, texture:TextureRegion, posX:Float, posY:Fl
 
   }
 
-  override def damage(source:Thing, amount:Integer): Unit = {
+  override def damage(source:Thing, amount:Float): Unit = {
 
   }
 }
